Move Session/Builder to their own file
